Initial commit of MATLAB work from 2012 grad class
About 2MB of files archived in this commit, originally from the
grad school class of Dr. Dragan Djurdjanovic from UT Austin.  It's
messy, but the file names state what data each .m-file was supposed
to calculate.  Note that for some files, they weren't meant to be
run as one giant process, but rather in chunks, in order to get
graphs and other data before it gets overwritten by the next part.